src/requirements.txt
Django==2.2.3

psycopg2-binary==2.8.3
xhtml2pdf==0.2.3

django-simple-history==2.8.0
